fix: 🐛 circular dep, fixture included in build
