+++ design-toolkit/components/src/components/query-builder/action-element.tsx
@@ -10,7 +10,9 @@
  * OF ANY KIND, either express or implied. See the License for the specific language
  * governing permissions and limitations under the License.
  */
+'use client';
 
+import 'client-only';
 import type { PressEvent } from '@react-types/shared';
 import { useCallback, useMemo } from 'react';
 import type { ActionProps } from 'react-querybuilder';
