+++ design-toolkit/components/src/components/query-builder/group.tsx
@@ -10,7 +10,9 @@
  * OF ANY KIND, either express or implied. See the License for the specific language
  * governing permissions and limitations under the License.
  */
+'use client';
 
+import 'client-only';
 import { cn } from '@/lib/utils';
 import { useMemo } from 'react';
 import {
