+++ design-toolkit/components/src/components/query-builder/index.tsx
@@ -9,27 +9,18 @@
  * OF ANY KIND, either express or implied. See the License for the specific language
  * governing permissions and limitations under the License.
  */
-
 'use client';
+
+import 'client-only';
 import { Delete, Duplicate, LockFill } from '@accelint/icons';
 import type { PressEvent } from '@react-types/shared';
-import 'client-only';
-import { type ReactElement, createContext, useCallback, useMemo } from 'react';
+import { createContext, useCallback, useMemo } from 'react';
 import {
   type ActionProps,
   type Classnames,
   type CombinatorSelectorProps,
-  type FullCombinator,
-  type FullField,
-  type FullOperator,
   QueryBuilder as RQBBuilder,
-  type Classnames as RQBClassnames,
-  type Field as RQBField,
-  type QueryBuilderProps as RQBProps,
-  type RuleGroupType as RQBRuleGroupType,
-  type ValueEditorProps,
 } from 'react-querybuilder';
-import type { LiteralUnion } from 'type-fest';
 import { cn } from '../../lib/utils';
 import { Button } from '../button';
 import { Icon } from '../icon';
@@ -43,67 +34,10 @@ import {
   RuleGroupHeaderComponent,
 } from './group';
 import { Rule } from './rule';
+import type { QueryBuilderContextType, QueryBuilderProps } from './types';
 import { getValidationResult, pressToMouseEvent } from './utils';
 import { ValueEditor } from './value-editor';
 import { ValueSelector } from './value-selector';
-
-export type RuleGroupType = RQBRuleGroupType;
-export type Field = RQBField;
-
-export type DefaultRQBProps = RQBProps<
-  RuleGroupType,
-  FullField,
-  FullOperator,
-  FullCombinator
->;
-
-// TODO: need to add multiselect back in when we have a compatible component
-export type QueryBuilderValueEditors = Record<
-  LiteralUnion<
-    Exclude<ValueEditorProps['type'], null | undefined | 'multiselect'>,
-    string
-  >,
-  (props: ValueEditorProps) => ReactElement
->;
-
-/**
- * Omitted props are currently unsupported functionality
- */
-export type QueryBuilderProps = Partial<
-  Omit<
-    DefaultRQBProps,
-    | 'showCombinatorsBetweenRules'
-    | 'independentCombinators'
-    | 'listsAsArrays'
-    | 'enableDragAndDrop'
-    | 'showNotToggle'
-    | 'showShiftActions'
-  > & {
-    orientation?: 'horizontal' | 'vertical';
-    showRuleLines?: boolean;
-  }
->;
-
-/**
- * Omitted classnames are for unsupported features
- */
-export type ClassNames = Omit<
-  RQBClassnames,
-  | 'betweenRules'
-  | 'branches'
-  | 'dndDragging'
-  | 'dndOver'
-  | 'dndCopy'
-  | 'dndGroup'
-  | 'dragHandle'
-  | 'shiftActions'
-  | 'notToggle'
->;
-
-export type QueryBuilderContextType = Pick<
-  QueryBuilderProps,
-  'showRuleLines' | 'orientation'
->;
 
 const operatorDescriptions: Record<string, string> = {
   AND: 'All rules below must be true for a match',
