+++ design-toolkit/components/src/components/query-builder/value-editor.tsx
@@ -10,7 +10,9 @@
  * OF ANY KIND, either express or implied. See the License for the specific language
  * governing permissions and limitations under the License.
  */
+'use client';
 
+import 'client-only';
 import { useCallback } from 'react';
 import { type ValueEditorProps, useValueEditor } from 'react-querybuilder';
 import { Checkbox } from '../checkbox';
@@ -19,8 +21,8 @@ import { Radio } from '../radio';
 import { Switch } from '../switch';
 import { TextAreaField } from '../text-area-field';
 import { TextField } from '../text-field';
-import type { QueryBuilderValueEditors } from './';
 import { multiValueOperators } from './constants';
+import type { QueryBuilderValueEditors } from './types';
 import { getValidationResult } from './utils';
 import { ValueSelector } from './value-selector';
